Handle inconsistent event syntax for params and text
Depending on the server, the target for part is sometimes given as one
of the params or as the trailing text. Handle both cases.

Display a privmsg differently depending on whether the target is the
channel or a nick.

Handle both cases in 333 RPL_TOPICWHOTIME where the time is given as a
param or trailing text.