Add support for encrypted ssl connections
For cl+ssl to work, the openssl library (libssl) has to be installed
on the system.

The connection can be encrypted by adding an appropriate port and
the keyword ssl to a server definition.

When SSL is enabled, the user mode +Z is set by the server, and the
numeric reply 671 +:rpl-whoissecure is returned by whois.