Add SASL PLAIN authentication for NickServ logins
Add basic support for IRCv3 capability negotiation (CAP LS, REQ, ACK),
currently only for SASL.

Add a function to encode the SASL PLAIN token in Base64.

SASL replaces the manual identification using /msg NickServ identify.

For SASL to work, it has to be explicitely enabled by passing the key
:login-method :sasl to the server object, because the default login
method is nil and the NickServ account and password have to be given.

SASL PLAIN transmits the password in Base64-encoded plaintext, so it is
useful to enable SSL encryption when logging.