Don't require msg as the first arg in define-event
Before the macro define-event required the name of the variable
that represents the parsed message object to be included in
every argument list.

But it turned out that we never need the object itself, we mostly
need only to access single slots like nick, params, etc.

So the message variable is removed from the argument list and can
be accessed as the macro-internal variable %msg%.

The arguments to define-event are now all accessors to %msg%.