Add an optional side window to display the buffer list
Add function and config option show-buffer-list to add a window
containing available buffers to the ui layout.

Bind it by default to F4.

Add the function update-buffer to redraw the buffer.

Add commands /show and /hide to toggle display of optional
windows, beginning with the buffers window.

Replace the separate calls to update parts of the ui with one
single function.