Add a way to toggle the topic, add an autojoin initarg
Allow /part to work without a channel argument, which
then leaves the current channel.

Add function and config variable show-topic-line to
toggle the display of the topic line.

Allow /show and /hide to also toggle the topic line.

Add the :autojoin initarg to connections, which takes
a comma-separated list of channels to join after login.

Rename function parse-posix-argv to handle-command-args

New and modified key bindings:

^A  - move to the start of the input line, instead of
      exiting the client.
F3  - toggle the display of the topic line
F7  - part the current channel
F12 - /exit the client